Add a app resource action lock
Startup scripts are not reliable during start and stop, so
we are better blocking the status calls during those actions.

This patch make an app resource acquire a resource-specific lock
on start, stop, status. If the action is status the lock is
released immediately.

Also remove the implicit rid from S* C* K* symlink support from
the resApp module, as it causes // start to be serialized.