Implement automodules
Previously, a moduleset referencing a module not available in
var/compliance/ was considered an error.

Now the agent sort (by name) the variables of the rulesets
embedded in the moduleset, and run the compliance object
associated with the variable class.

If the action is 'check' or 'fixable', the loop over variables
continue on error.  If the action is 'fix', the loop over
variables breaks on first error.

With this patch, codeless compliance designs are possible and
can handle most needs. More complex designs can still implement
a module.

Compliance runs output changes:

* RCODE is also renamed to STATUS, and uncolored. The status value
is still colored, but printed in a human readable form (ok, nok,
n/a).

* DURATION is removed, to unclutter the output.