Fix clustertype setting
All services were considered failover, whatever the cluster_type
defined, because the Svc class now sets "failover" as a
placeholder value, and the service builder still considered
that the presence of the Svc::clustertype property means it
should not be reset.

With this patch, the builder now resets the Svc::clustertype.

One consequence of this issue was that the sync were allowed
from flex secondaries to the flex primary node.