Remove the "nodemgr provision" action, add "nodemgr array"
nodemgr provision was too restrictive, as limited to disk adds
and induced a useless rebound in a prov*.py module.

Implement instead a "nodemgr array" action wrapping the rc<array>
modules, which now reexport to the agent the low-level, array
specific, actions.

"nodemgr array" commands are meant to be called from the
collector forms, which implement the client-specific policies.