Always sort svcmon output by svcname
"svcmon -c [-v]" already had its output sorted, but the parallel
execution of "status eval + status print" when svcmon was
executed without "-c" made the services to be displayed in a
random order, as soon as evaluated.

This patch stops doing the print in the worker. The worker is
now only responsible for refreshing the status caches. Moreover,
the parallelization is disabled if --refresh is not set, as
fetching the status from the caches is fast enough to not
justify the multiprocessing.

Overall, this patch simplifies the code a lot, as we no longer
need a queue to gather status to feed to the collector in a
single request, nor a lock to prevent worker prints from
interleaving.