Don't stack when the service configuration file is corrupted
beyond python ConfigParser capabilities.

For example, when a configuration does not begin with a section.

With this patch, the following result is emitted:

$ sudo testsvc edit config --recover
15:36:25,043 testsvc    ERROR   error parsing /opt/opensvc/tmp/testsvc.conf.tmp
15:36:25,045 testsvc    ERROR   the new configuration causes the following build error: File contains no section headers.
file: /opt/opensvc/tmp/testsvc.conf.tmp, line: 1
u'svc-lxc-centos6\n'
your changes were not applied because of the errors reported above. you can use the edit config command with --recover to try to fix your changes or with --discard to restart from the live config