Fix the "pool#<n>" resources backward compat
pool# resources have a special treatment as the zpool# resource
are also accepted: pool# are translated to zpool#, but the
logic was bogus as we still evicted pool# resources in the test
to evict from add_disk() sections of different resource family.

Fix that, and optimize the loop in disk_add(), moving outside
a code block producing an invariant result.