Support pointing a docker service rid in a docker ip resource
In this case, the first local container found is chosen to plumb the ip
into. As a consequence, the pointed docker service should be configured
for a single replica.

With this feature, one can declare docker services without any publishings.
They share at least one overlay network in the swarm with the container
hosting the public ip. This container can thus act as a reverse proxy for
the swarm services.

If the swarm orchestration moves the container, setting the docker ip
resource monitored and with restart>0, opensvc will automatically plumb
the public in the new container at its new location.