Mangle the docker ip resource status linked to a docker service
If the ip does not ping anywhere in the swarm, change the status from down to
stdby down, else if the ip is not up change to stdby up.

These tweaks are necessary for the resource restart to trigger. If we let the
ip resource status evaluate as 'down', the resource restart algo will consider
it not legitimate to start the resource.

With this patch a ip switchover in a swarm rolls up like that:

node1   node2
=====   =====
up      stdby up
(death) stdby down => restart
        up

The resource monitor scheduled task is responsible for the resource restart,
so a delay of ~1min applies.