Don't log task run_requires constraints not met as error if scheduled
The run_requires parameter is meant to allow tasks run bypass
conditions, so not meeting the constraints is not to be considered
an error when the run is tried by the scheduler.

But if the user explicitely asks for a run, not meeting the
constraints still is an error.

Example:

$ sudo testmd run --cron --rid task#1
15:44:33,752 task#1      INFO    requires on resource fs#4 in state up or stdby up, current state down
15:44:33,753 task#1      INFO    continue action on resource set

$ sudo testmd run --rid task#1
15:44:42,210 task#1      ERROR   requires on resource fs#4 in state up or stdby up, current state down