Report fs.dir resource "down" if the directory does not exist
Instead of "n/a + warn resource log".

This patch allows a service to use a fs.dir resource as a
the simplest service "started" state flag, basing the state
on the existance of a directory (usually the directory
hosting the app scripts).

For example, an encap service can now evaluate its state to down
if it has only app resources, plus a fs.dir resource.