Separate OSVC_SHUTDOWN_OPTS from OSVC_BOOT_OPTS
Users settings a custom OSVC_BOOT_OPTS can set there an
option (--disable-rollback for example) valid for the boot
action but not valid for the shutdown actions.

This patch introduces in all startup script a OSVC_SHUTDOWN_OPTS
with a default value set to --parallel, like OSVC_BOOT_OPTS.

This new variable is overridable through the /etc/default/opensvc
or equivallent for each os. As such the new parameter is now
also included in the /etc/default/opensvc template installed
with the agent.