Fix warning on docker resources with docker v1.12.6 (rhel)
A resource can reference the image by its short symbolic name,
like "ubuntu:16.04". In this case the pull downloads from
docker.io, and some versions of docker append docker.io/ before
the name, breaking the algo looking for the image id.

Starting with v13 we use image inspect to do the lookup, but we
need a fix for the affected docker version range.

This patch just drops comparing the name part of the image and
just compares the tag. This is simpler and valid because we know
the "docker images <run_image>" output can only contain tags of
the looked up image.