Allow Resource::label as a lazy prop
* Catch AttributeError in Resource init when setting the label
property, as if the child class defines label as a lazy, this
exception would be raised.

* Move the label override after parent Resource::__init__() in
all resource drivers setting it before.

* Define label as a lazy in the docker resource container, as an
optimization of the service build. svcmon, for one does not need
the label, so there no need to always format it.

* Also set the docker resources service_id and container_id as
lazy props, so we can remove the on_add() method to further
optimize the build.