Optimizations
This batch of optimizations is driven by an analysis of a
"strace -e execve svcmon --refresh" run.

Add more cached functions:

* disk.raw: raw.rdevs_t, raw.modprobe, raw.list
* sync.rsync: rsync.version

Add more lazy props:

* disk.raw: devs_t, rdevs_t, raws
* rsync: full_options

Execute system binaries through their fullpath. The rcEnv.syspath
Storage() is used to store the per-os reference paths. This is
good for security and is a micro-optimization as the libc doesn't
need to try-loop all paths in PATH.

- syspaths.blkid
- syspaths.dmsetup
- syspaths.ip
- syspaths.losetup
- syspaths.lsmod
- syspaths.lvs
- syspaths.mount
- syspaths.multipath
- syspaths.multipathd
- syspaths.umount
- syspaths.vgs
- syspaths.zfs
- syspaths.zpool

Add a cache to the rcMounts::get_dir_src_dev(dev) method.