Add a DEFAULT.pre_monitor_action trigger
A script to execute before the monitor_action. For example, if the
monitor_action is set to freezestop, the script can decide to crash the
server if it detects a situation were the freezestop can not succeed
(ex. fs can not be umounted with a dead storage array).

This patch also exposes the "toc" service action.