Zpool driver enhancement
Teach the zpool driver to look into the new wwn_map.<rid> state
files sent by the service holder node to its slave.

If a device file found in the disklist cache sent by the holder
node is not present on the standby node, try and see if the
device is not remappable to another device.

This case happens when the holder node uses R1 SRDF volumes,
while the standby node uses R2. R1 and R2 having different wwn,
the device needs remapping for disklist (and scsi reservation
status and start) to work.

The SRDF sync driver now takes care of replicating the mapping
amongst cluster nodes.