Read the disk mapping file both ways
The mapping file can be generated R1 or R2 side, and R1 sends it
to the R2. The local disks being always located in the left-most
mapping column, we need to index the mapping both ways.

This fixes scsireserv status refresh when the co-resource is a
zfs on R2 disks.