Fix the --refresh flag unduly set in encap commands
For example, when trying to create (or update) the service config
in the container from a scheduler run, the refresh flag is set
but we don't want to set --refresh with a 'create' action as it
is an invalid combination.

This patch removes the blind relay of options.refresh, and let
the encap_cmd() callers decide when they want it set.