Add the service DEFAULT.affinity keyword
Teach the daemon monitor not to start a service if the list of
service defined as affine are not already start on the same
node.

Remove the --ignore-affinity svcmgr option, and the affinity
safety nets in the Svc class: Now that the daemon is responsible
of enforcing affinity and anti-affinity, we can let the admins
override when bypassing the daemon.