Implement cluster-wide stop and start ops
* Add a global_expect service monitor attribute
* Rename expect to local_expect
* Fix the unset of service monitor attributes
* Rename the smon expect status: up => started
* Split service_orchestrator() into _manual() and _auto()
* Discard frozen instances from the placement leader election
* Log the ready state remaining time in each monitor loop
* Rename get_service_status*() to get_global_service_status*()
* Drop the --remote svcmgr option

Stop behaviour:

1/ svcmgr sends a set_service_monitor action its local daemon,
with global_expect="stopped"
2/ the monitor loop enters service_orchestrator_manual() instead
of service_orchestrator_auto() because of global_expect being set
3/ service_orchestrator_manual() freezes the local instance if not
already frozen, and stops it if not already stopped.
4/ the hb data is updated and the heartbeats propagate the
global_expect value
5/ on a remote daemon, the foreign global_expect is detected and
adopted locally by merge_hb_data() if the local service instance
is not already aligned with the global expect.
In the former cas, steps 2 to 4 are applied.
6/ when the service instance is actually stopped, the global_expect
is dropped.

Start behaviour:

1/ svcmgr sends a set_service_monitor action its local daemon,
with global_expect="started"
2/ the monitor loop enters service_orchestrator_manual() instead
of service_orchestrator_auto() because of global_expect being set
3/ service_orchestrator_manual() thaw the local instance if not
already thawed and executes service_orchestrator_auto() to let
the placement policy and constraints apply.
4/ the hb data is updated and the heartbeats propagate the
global_expect value
5/ on a remote daemon, the foreign global_expect is detected and
adopted locally by merge_hb_data() if the local service instance
is not already aligned with the global expect.
In the former cas, steps 2 to 4 are applied.
6/ when the service instance is actually started, the global_expect
is dropped.