Add service-level triggers
pre, post, blocking_pre, blocking_post for most actions are now
supported in the DEFAULT section and interpreted as global
service action triggers.

This is the facility to use to replace the openha startup script
mecanism.

This patch also moves the action_triggers() from the Resource
class to rcUtilities, so we can use it in both Svc and Resource
classes.

It also introduces a use_default boolean flag to all service
configuration get helpers. triggers are the first example of
keywords never inheriting their value from the DEFAULT section.

The triggers discovery is removed from the service build, and
deferred to the moment it is actually needed.