osvcd fixes and enhancements
* Fix the service config validation, broken by the move of
conf_get* helpers to the Svc class

* Add the "nodemgr daemon join --secret <secret> --node <node>"
command, which adds the local nodename to the remote cluster
nodes list and aligns the local cluster and heartbeat sections
to those of the joined node.

* Reload node config in all osvcd threads. only the daemon scans
for node.conf changes, but now it notifies all threads to refresh
their cache. This patch notably enables hot secret changes.

* Fix python3 and LANG=C service config file management. Factorize
the read_cf() helper in rcUtilities as it has all the python2/3
abstraction.

* Do the node.conf write in the set() codepath in utf8, like all
other codepaths.

* Use rcConfigParser.RawConfigParser in osvcd instead of the
python class. To benefit from utf8 python2/3 abstractions.

* Care about the creation of the "cluster" section in node.conf
when we generate a new secret.

* Better error logging of decrypt issues

* Add nodename to the cluster.nodes list when receiving a message
encrypted with our secret. This allows a new node to join any
other node in the cluster: the peers will discover the new node
through the mcast or ucast rx threads.