Fix the container abort start routine when hostname != name
Use the container hostname in the abort start check. When
hostname is different than name, and the name actually ping
the service start would be blocked. But the 'name' can
resolve to a different IP than hostname and this IP can be
up on an unrelated infrastructure component.

Pinging the hostname is really what we want.