Add a disk heartbeat to osvcd
Tested with directio and mmap on Linux for now.

The definition is:

[hb#3]
type = disk
dev = /dev/mapper/36589cfc000000fb33de590d7e9e2bfec

The first 4M of the disk are reserved for the node slots index.
A page (4k) per slot index, ie a 1024 nodes limit.
The rest of the disk is segmented into 1M slots, where each node will
write its encrypted message.

A slot index contains the nodename.
Upon start-up, the tx thread searches for its nodename in the index.
If not found it reserves one, writing its nodename in the slot meta.

The rx thread re-reads the meta until it finds an empty meta slot, building
the list of nodes (cross-checked with the cluster.nodes setting) to read
the data in their respective slots.

The data contains the nodename too, so the reader is able to alert when
a node does write in its excepted slot.