Fetch service configs advertized by peers
* Add the service "scope" key to the hb payload, alongside
services config cksum and mtime. The scope is the concatenated,
deduplicated list of nodes and drpnodes.

* Fetch and install a service config from the most up to date
peer when the node is a member of the "scope".

With this patch, deleting the local service instance without
previously remove the node from the nodes or drpnodes list will
lead to its reinstallation.