Fix service print actions returning 1 instead of 0
Fix the Node::need_aggregate() method to return False when the action
starts with "print" but the selected formatter is not json.

Only json data can be aggregated.

This issue caused "print config mtime" to return 1, and its outed data
to be discarded, which in turn broke the master/slave service config
file synchronization automation.