Add a diff to the 'svcmgr edit config' error message
The svcmgr 'edit config' action can send the following message:

"service is already being edited. Set --discard to edit from the current
configuration, or --recover to open the unapplied config"

In this situation, the user may want to see the diff to choose between
--discard and --recover.

This patch preprends a diff to the same message. And display the message
to stderr instead of the service log.

Example:

--- /opt/opensvc/etc/pridns.conf	2017-07-05 18:10:26.717698263 +0200
+++ /opt/opensvc/tmp/pridns.conf.tmp	2017-07-21 09:40:36.157973276 +0200
@@ -10,7 +10,6 @@
 gateway = {env.gateway}
 container_rid = container#0
 tags = docker
-optional = true
 dns_update = true

 [disk#0]

/opt/opensvc/tmp/pridns.conf.tmp exists:
service is already being edited. Set --discard to edit from the current
configuration, or --recover to open the unapplied config