builder/svcDict integration enhancements
* implement candidates validation for list, tuple and set values. The
value is considered valid if all members of the value are in the
candidates.

* remove all keyword value spliting from the build. flag them as
convert=list or convert=set instead, so the spliting is done by
Svc::conf_get()

* support the list values in Svc::_handle_reference(), we previously
only received strings here

* remove all keyword value cast to bool and int in the builder. set and use
the keyword converter instead.

* remove optional_on, monitor_on, enable_on and disable_on keywords.
disable, monitor and optional scoping is enough

* add the 'inheritance' keyword property. possible values are 'leaf',
'head', 'head > left', 'leaf > head'. the default is 'leaf > head'. only
the 'disable' keyword uses the 'head > leaf' behviour so far.

* add the 'scope_order' keyword property. possible values are 'specific
> generic', 'generic > specific', 'head: generic > specific'.
the default is 'specific > generic'.

* update docs