Clarify a long-standing ambiguity with devs/disks meaning
devlist/disklist are not precise enough a foundation to build upon.

Replace them with:

* exposed_devs
  devices a resource exposes
  resources-specific.

* sub_devs
  devices a resource is layered above
  resources-specific.

* base_devs
  devices at the base of the device tree a resource uses.
  this method is common to all resources, based on a devtree lookup.

The "print devlist" and "print disklist" are replaced by

* print exposed devs [<rid selectors>]
* print sub devs [<rid selectors>]
* print base devs [<rid selectors>]
* print devs [<rid selectors>]

This last command aggregates all of the precedent information.
The outputs are rendered using the forest module, or in json format.

The scsi reservation drivers now use peer resource sub_devs().

The devtree is now cached as a lazy property of the Node object.