Don't use mmap on the hb disk
This cause python crash on BUS ERROR when the hb disk returns I/O errors,
which what happens when all paths to the disk go offline.

Instead, use a mmap-ed 4k buffer for slot metadata and a 4M one for slot
data, and use fdopen() on the O_DIRECT-opened fd to satisfy memory
alignment constraints of the file object readinto() and write() methods.

With this patch, I/O errors are logged, and properly retried.