app resource changes and enhancements
Formerly an app resource was defined as:

[app#0]
script = /bin/true
check = 50
stop = 50
start = 50
info = 50

on start, "<script> start" is executed.
on stop, "<script> stop" is executed.
on push resinfo, "<script> info" is executed.
on status evaluation, "<script> status" is executed.

The start,stop,check,info value was used for ordering app resources. If one
of these keywords was not set, the resource was skipped in the corresponding
action.

This patch address a bunch of shortcomings:

1/ the ordering method is not standard, for no benefit.
2/ <script> being common for all actions, it makes it hard to reuse
existing scripts, accepting parameters not aligned with opensvc requirements.

The new syntax is:

[app#0]
script = /bin/true
start = /bin/true custo start
stop = true
check = 50

This configlet has the following behaviour:

on start, "<script> start" is executed.
on stop, "<stop>" is executed, being more precise than <script>.
on push resinfo, do nothing (info is not set to a bool nor a int nor a command)
on status evaluation, "<script> status" is executed (compat code to ease upgrade)

The standard resource ordering applies here (rid is used as the sort key).
In the former configlet, app#<start> would be used as the sort key.