provision action behaviour change
Up to this patch, the provision instance action was expected to
leave the instance in up status.

This behaviour is no longer desirable in cluster-wide service
provisioning. As is would create conflicts.

This patch implements:

* ip, share and app resource no longer start on provision
* at the end of a provision action, do a rollback
* keep rollback from stopping always on resources

A provisioned instance is thus left in stdby up or down status,
ready for orchestration.