Implement prov/unprov/purge sequencing
Service with shared resources must unprovision on a single node
first, then unprovision on the other nodes, so the shared
resources don't get destroyed simultaneously from different nodes.

This patch implement the election of the node that will
unprovision first and the wait mecanism.

The elected node is the best ranked peer with a up instance, or
as a fallback, the best ranked peer.

This patch also prevent setting the local_expect to "started"
if a global_expect is set or the smon status is not "idle".