Don't raise on linux.lvm vgs pvs and lvs commands error
These are interrogation commands, but they still can return a not-zero code.
One case is the "invalid metadata" autorepair error because one pv is read
only.

The stdout is valid though, so don't be radical and accept parsing it, even
when ret != 0.

If not, all vg resources of all services will be considered undef, which
would cause monitor_action execution on services configured to monitor such
vg resources.