Drop the stonith drivers and "svcmgr stonith" action
This stonith methods were limited to 2-nodes clusters. Stonith will be
reimplemented as a node action, and executed by the daemon monitor.

The service still needs to be able to declare it mandates a stonith. To
this end, a new DEFAULT.stonith boolean keyword is added. It defaults to
false.