Add "nodemgr stonith --node <nodename>"
In the node.conf, users can define a section like this for each cluster node:

[stonith#<nodename>]
cmd = /bin/true

The "nodemgr stonith --node <nodename>" command would then execute /bin/true.

This is similar to the stonith.callout resource driver. The stonith.ilo
driver is not forward ported, and users should use fencing commands
provided by the standard fencing packages available.

This patch also add a "stonith" boolean property to the instance status
data used by the daemon. This property is set to True if the instance is
avail up and the DEFAULT.clustertype is failover and DEFAULT.stonith is True.
Else it is unset.

This property is thus a "shoot me" indicator that the daemon will use to
decide if a stonith action is mandated.