systemd integration fixes
* "nodemgr daemon shutdown" requalifed as a stop always returned an error,
even when the job was done right

* "nodemgr daemon restart" when the running daemon was started manually
(systemd reporting it 'inactive' when it's actually running), failed.
In this case systemd thinks it can bypass the stop and do the start right
away, which fails as a daemon is already running.

* "nodemgr daemon stop" did not use the systemctl command, leading to
systemd restarting the daemon on failure.

* "nodemgr daemon start" did not use the systemctl command, leading to
systemd thinking the daemon is inactive, when it's actually active.