container.lxd enhancements
* fix the provisioner/unprovisioner methods incorrectly named
start/stop. This caused the unprovision action not to stop the
container before delete.

* don't run "lxc launch" with stdin closed, as this causes
confusion for lxd

* wait for the container to appear as "Running" before rollbacking
the start caused by the "lxc launch" command. If the container is
stopped too soon, a hang happens frequently.

* fix options unduly set on "lxc delete", and the command being
run twice

* make is_up() and lxc_info() take a peer nodename as argument
and use that to implement a is_up_on(nodename) method and a
lxd-specific abort_start() method which don't require the
container ip to be identified and pingable.

* on zfs, recreate the @readonly image snapshot after promoting it,
so lxd is ready to launch another container based on this image.