Stop playing tricks with the lxd handled zpool
We were trying hard to make the recursive zfs send/recv work,
but it just not possible with container datasets being clones
of their image. The remote node has the same image stored as an
independent dataset, so the send/recv can not maintain the
dataset-to-clone relationship.

So, users need to know that sync.zfs resources need recursive=false
when synchronizing the a container root dataset.