sync.zfs fixes
Try harder not to run the pre_sync() of sync.zfs resources when
the situation is not opportune.

The provision situation is notably inopportune to run the pre_sync
action, where a @tosend snap can be created on the nodes not
placement leaders ... which cause the first legit sync on the
leader to try an incremental instead of a full.