Cluster-wide provisioning fix
Wait for all live nodes to have received the new service config
file before proceeding to the provision action.

If not, the node receiving the provision request will start
provision as leader first, thinking it is the only candidate node
for the service. Then as the real leader receives the config, it
will provision too as leader.