Avoid unwanted takeover on long-dead node rejoin
This new algo in the monitor loop is only called during the
grace period.

It freezes the local services instances for services that have
a live remote instance frozen.

This prevents a node rejoining the cluster from taking over services
that where frozen and stopped on peers while we were not alive.