Move status and config push to collector into a new daemon thread
The daemon thread is named collector. It loops at 5 minutes max interval,
and is waken up by the monitor thread at the end of its loop.

The service json status now embeds a checksum of all values expect dates, so
this checksum changes are interpreted as a need to push the instance status
to the collector.

Likewise, the config file checksum changes are interpreted as a need to push
the config file and service instance properties to the collector.

Only the first live node of a cluster is allowed to send status to the
collector. This is the speaker node of the cluster. This speaker sends
status even for services it has no local instance of.

Now svcmon on any node shows all cluster services, even if the node has no
local instance.

When the collector thread doesn't detect service status changes, it sends a
ping message to collector. Else the full daemon thread status is sent, with
a digest of the detected changes as a list like ["svc1", "svc1@node1"]. In
this example, only the svc1 instance on node1 has changed.

Remove the push_config service task, and the autopush for the service builder.
Reinstate a push_encap_config() call in the _encap_cmd() codepath.

Rename the push_service_status service task to "status", has it is no longer
responsible for pushing.

Add new keys to json status:
+ placement
+ flex_min_nodes
+ flex_max_nodes
+ topology
+ parents
+ children
+ enslave_children