Add the "svcmgr abort" action
As "svcmgr clear" just clears the service monitor status, the target state is
preserved and the daemon will retry to reach this target asap ... and maybe
run into the same problem that mandated a "clear".

We need a way to break the loop.

The new action "abort", resets the target state and also clear status.