Evaluate a sync status to n/a if the service is not avail-up cluster-wide
sync drivers refuse to run on non-avail-up instances, so when no instance
is avail-up the sync_max_delay ends up overflowed and the sync resource status
is degraded to warn.

This is wrong, as the sync'ed data do not change when the service is not
avail-up. So prefer a n/a status in this case, which is easy now that the
daemons provide the cluster-wide service status.