Notify the daemon of local actions begin and end
So the service monitor data is always up-to-date, which benefits
to:

* svcmon watchers, notified immediately of actions not initiated
by the daemon
* monitor orchestrations, which like TOC and resource restart can
back off during actions not initiated by the daemon

This fixes the giveback action unduly triggering a TOC because the
daemon was not aware of the instance stop.

This patch also adds a safety net in the TOC codepath, refusing
to proceed if the instance is frozen.