Move the instance monitor status handling to the CRM
Upon instance action, verify the status is idle.

Don't set the status from the daemon, as it would block the CRM action.

With this patch, it is no longer possible to start a sync while starting,
stopping, deleting, freezing, thawing the instance. And it is no longer
allowed to start, stop, delete, ... while syncing.