Execute the resourcesets pre/post action at their expected step
Currently, the post_action resourceset method executes at the end of the
overall action, but it is expected to execute just after the resourcesets
of the same class.

This caused the dockerd stop to be triggered too late, after the fs umount
caused them to be killed through fuser -k.