Implement an arbitrators-based quorum
New node.conf parameters control that behaviour.

[cluster]
quorum = true

If set to true, a split segment of the cluster commit will commit
suicide if the sum of the votes of the visible nodes + the vote of
a reachable arbitrator is less than half the number of nodes in the
cluster.

The default is False.

If set to true, please set at least 2 arbitrators so you can rolling
upgrade the opensvc daemons.

An arbitrator is a opensvc node (running the usual osvc daemon) this
cluster nodes can ask for a vote when the cluster is split.

Arbitrators are tried in sequence, the first reachable arbitrator
gives a vote. In case of a real split, all arbitrators are expected to
be unreachable from the lost segment. At least one of them is
expected to be reachable from the surviving segment.

Arbitrators of a cluster must thus be located close enough to each
other, so a subset of arbitrators can't be reachable from a split
cluster segment, while another subset of arbitrators is reachable
from the other split cluster segment. But not close enough so they can
all fail together. Usually, this can be interpreted as: same site,
not same rack and power lines.

Arbitrators usually don't run services, even though they could, as their
secret might be known by multiple clusters of different responsibles.

Arbitrators can be tested using "nodemgr ping --node <arbitrator name>"

[arbitrator#0]
name = node3
secret = 12331241421413413