Keep the scheduler thread from running the same command twice
Keep a set of running command signatures, and refuse to run a subprocess for
a command already running.

Also embed the running command set length in the thread status data.

This protection avoids piling subprocess in situations where the subprocess
never terminate, which is what happens when they get stuck in D-state.