Service json status format changes
Remove keys from the json status when their value is default.

At top-level:

* frozen
* children
* parents
* children
* enslave_children
* constraints

At resource-level:

* disable
* optional
* standby
* encap
* monitor
* log
* info

Move the group status to a "status_group" subkey.

Embed resource logs as a list of strings instead of a string.

Bump the compat version to 5.

And remove the Resource::status_quad() method, misnamed and no longer useful.

The collector must be updated to accept this new, more compact, format.