Add a reboot.once boolean keyword to node.conf
If once is set to false, do not remove the reboot flag before rebooting,
so that the node is ready to reboot again in the next allowed timerange.
This setup is needed to enforce a periodic reboot, with a patching script
hooked as a pre trigger for example.

If not set, or set to true, the reboot flag is removed before reboot, and
a 'nodemgr schedule reboot' is needed to rearm.