Catch possible KeyErrors in the resource provisioned state merging
Can happen when an invalid service config is propagated and
trashed at on the receiver end, and the config is fixed on the
sender.

At this point, there is a window were the sender will include
resource provisioned states to merge on the receiver, which still
does not know the service resources.