Faster monitor proc janitoring loop and ready stealing fix
When the monitor thread queues a process it the next janitoring,
hence the callbacks execution, will only happen on the next
loop iteration. That it about 5 seconds in general.

Shorten the janitoring delay to 1 second, and run the rest
of the monitor loop only if the short loop ran 5 times or if
the proc janitor reports some process are terminated.

This patch also allow a local service instance to enter the
ready state even if another node is ready: the placement policy
will determine which one should back off.

On the other hand, disallow entering the ready state if a peer
instance is in provisioning. Let the peer finish before we
go ready.