Get rid of the per-type resourceset
This patch merges the per-type resourcesets into a default resourceset.

Resources in this resourceset are ordered by rid, expect the *pr resourceset
whose position is forced just before their peer resource.

Now the subsets ordering and default resource set ordering are thus coherent.

Be aware that you now have to name resources in the default subset so a
alphanum sort sort them in the expected start order. You can no longer rely
on the agent to force an order between drivers of the same family.

As a benefit most setups using subsets just for the sake of ordering can now
move to a simpler subset-less configuration.