Fix delete loop when deleting a service with nodes={clusternodes}
In this case the "set --kw nodes-=<node>" doesn't remove the node, so
the peers keep fetching the configuration file, thinking it's a new
service.

This patch makes the listener return a specific value (2) when a peer
is trying to fetch the config of a service in "deleting" state. This
return value is used by the client to log an info and skip.