Relax the sync.symsrdfs presync export condition
Trust the avail status instead of the overall status to determine if we
are master, and thus allow re-export of the symdg configuration.

This avoids missing the symdg reconfiguration when a swithover to R2-side
lead to a situtation where the instance is avail up, but with warnings or
with optional resources down.