Support scoping the node.conf cluster.name and cluster.secret keywords
And impersonate the remote node when getting those keys, preparing a
remote_service_action() to submit to the remote daemon.

This is necessary when the remote is not a member of the sender's cluster,
which is what happens when a node tries to trigger a postsync() on a
drp node after a sync_drp().