Better handling of the cluster upgrades
Instead of freezing the whole cluster and leaving the thaw at discretion of
the operator, implement the following:

* Add a "upgrade" node monitor state, which like "maintenance", persists
service status in the peers but with no expiration.

* Add support for the OPENSVC_AGENT_UPGRADE env variable on daemon stop and
daemon start action. If set, switch node state to "upgrade" instead of
"maintenance".

* On startup with OPENSVC_AGENT_UPGRADE set, the monitor thread freezes the node
and remembers it has to thaw when the cluster is complete. If the node was
already frozen, leave it alone.

With these policies a cluster should behave when all nodes are upgraded in
a random order or near-simultaneously.

If a node agent fails to restart, the cluster-complete state won't be reached
and the successfully upgraded nodes will stay frozen, and the not-upgraded-at-all
nodes will preserve forever the services data from the failed node. This
situation needs admin arbitration, but do not cause service disruption for
end-users.