Add a powerdns remote backend thread to osvcd
This thread serves through <pathvar>/dns/pdns.sock. It is started only if the
cluster.dns keyword is set.

For a start, this backend serves soa, a and ptr records:

* each service is a sub-zone of the cluster name
* each ipaddr in a service is served when querying "A <svcname>.<clustername>"
* if a container resource has a hostname, serve a
<hostname>.<svcname>.<clustername>" for the ipaddr of the container
* if an ip resource sets dns_name_suffix, the record becomes
<hostname><dns name suffix>.<svcname>.<clustername>
* scaler sub-services are transparent wrt the dns scheme

Add new node keyword:

* cluster.dns = <ip1> <ip2>

Add new references:

* {dns} resolved to the cluster.dns ip list
* {dnsnodes} resolved to the cluster nodes having cluster.dns ips
* {dnsuxsockd} resolved to the dir hosting the pdns.sock socket
* {dnsuxsock} resolved to the pdns.sock socket full path

Service status data changes:

* Remove the "enslaved_children" boolean key
* Add the "scaler_slave" boolean key
* Children no longer include the scaler slaves

If a docker container run_args does not specifiy --dns and cluster.dns is set,
use the cluster dns ips, and set a search with <svcname>.<clustername> and
<clustername>.